Moved the testes to the file test, removed unnecessary tests and updated code
